run intellij code cleanup on bufr, cdm, grib, netcdf4, udunits, ui, visadCdm

+++ visad/src/main/java/ucar/nc2/iosp/gempak/GempakStation.java
@@ -449,7 +449,7 @@ public class GempakStation implements Station {
   public String getWmoId() {
     String wmoID = "";
     if (!(stnm == GempakConstants.IMISSD)) {
-      wmoID = String.valueOf((int) (stnm / 10));
+      wmoID = String.valueOf(stnm / 10);
     }
     return wmoID;
   }
